Start work on configuration and basic functions
The architectural ideas here are simple: the core is not responsible
for the configuration file, nor does it have any knowledge of a
configuration file.

The frontend will parse its own configuration file, tell the core what
it found, and if the core has a problem with the values of a given
configuration, it will report that issue back to the frontend.

Signed-off-by: dgz <dgz@dgz.me>